recompiler: support for 64-bit enabled REM library on 32-bit hosts (disabled by USE_VBOXREM64 in Makefile.kmk now)
svn:sync-xref-src-repo-rev: r40690

+++ src/recompiler_new/VBoxRecompiler.c
@@ -1,4 +1,4 @@
-/* $Id: VBoxRecompiler.c 15015 2008-12-05 08:44:26Z noreply@oracle.com $ */
+/* $Id: VBoxRecompiler.c 15272 2008-12-10 18:47:19Z noreply@oracle.com $ */
 /** @file
  * VBox Recompiler - QEMU.
  */
@@ -3492,7 +3492,7 @@ bool remR3DisasBlock(CPUState *env, int f32BitCode, int nrInstructions, char *ps
     /*
      * Disassemble.
      */
-    off = env->eip - (RTGCUINTPTR)pvPC;
+    off = env->eip - (RTGCUINTPTR)(uintptr_t)pvPC;
     Cpu.mode = f32BitCode ? CPUMODE_32BIT : CPUMODE_16BIT;
     Cpu.pfnReadBytes = NULL;            /** @todo make cs:eip reader for the disassembler. */
     //Cpu.dwUserData[0] = (uintptr_t)pVM;
@@ -3593,7 +3593,7 @@ bool remR3DisasInstr(CPUState *env, int f32BitCode, char *pszPrefix)
     /*
      * Disassemble.
      */
-    off = env->eip - (RTGCUINTPTR)pvPC;
+    off = env->eip - (RTGCUINTPTR)(uintptr_t)pvPC;
     Cpu.mode = f32BitCode ? CPUMODE_32BIT : CPUMODE_16BIT;
     Cpu.pfnReadBytes = NULL;            /** @todo make cs:eip reader for the disassembler. */
     //Cpu.dwUserData[0] = (uintptr_t)pVM;
